requirements.txt
bcrypt==4.2.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
contourpy==1.3.1
cryptography==43.0.0
cycler==0.12.1
fonttools==4.56.0
google-auth==2.33.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
httplib2==0.22.0
idna==3.7
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.0.1
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.2
pandas==2.2.2
paramiko==3.4.0
pillow==11.1.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyNaCl==1.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
seaborn==0.13.2
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
